% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{academic_awards}
\alias{academic_awards}
\title{Simulated Sesame Street Data}
\format{A data frame with 200 rows and 4 variables.}
\usage{
data(academic_awards)
}
\description{
Simulated dataset from
\url{https://stats.idre.ucla.edu/sas/dae/poisson-regression/}.
}
\details{
#' this example, , ,
and prog is a categorical predictor variable with three levels indicating the
type of program in which the students were enrolled. It is coded as 1 =
\"General\", 2 = \"Academic\" and 3 = \"Vocational\".#'

\itemize{
  \item{"num_awards"}{Outcome variable; indicates the number of
awards earned by students at a high school in a year}
  \item{"math"}{Continuous predictor variable; represents students' scores on
  their math final exam}
  \item{"prog"}{Categorical predictor variable with three levels, indicating
  the type of program in which the students were enrolled. It is coded as 1 =
\"General\", 2 = \"Academic\" and 3 = \"Vocational\".}
}
}
\keyword{datasets}
