% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorica.R
\name{compare_hypotheses}
\alias{compare_hypotheses}
\title{Evaluate Informative Hypotheses}
\usage{
compare_hypotheses(object, ..., iter = 1e+05)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{iter}{PARAM_DESCRIPTION, Default: 1e+05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
academic_awards <- within(academic_awards, {
  prog <- factor(prog, levels = 1:3, labels = c("General", "Academic",
                                                "Vocational") ) } )
  zmath <- scale(academic_awards$math)
  #Model fitting
  model <- glm(num_awards ~ prog + zmath + prog * zmath, family = "poisson",
               data = academic_awards)
               new_gor <- gorica(x = model, hypothesis = "zmath=0&zmath+
               progAcademic:zmath=0&zmath+progVocational:zmath=0;
               progAcademic:zmath > progVocational:zmath&
               progVocational:zmath>0; progVocational:zmath<0")
}
}
