% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hox_2010.R
\docType{data}
\name{hox_2010}
\alias{hox_2010}
\title{Sesame Street data based on Hox (2010)}
\format{
A data frame with 2000 rows and 6 variables.
}
\usage{
data(hox_2010)
}
\description{
Synthetic data based on Hox (2010, p. 16).
}
\details{
\tabular{lll}{
  \strong{pupil} \tab \code{integer} \tab Pupil ID\cr
  \strong{school} \tab \code{integer} \tab School ID\cr
  \strong{popular} \tab \code{numeric} \tab Popularity scores,
  standardized\cr
  \strong{sex} \tab \code{factor} \tab Pupil sex\cr
  \strong{text} \tab \code{numeric} \tab Teacher experience, standardized\cr
  \strong{teachpop} \tab \code{integer} \tab ??
}
}
\references{
Hox, J. J. (2010). Multilevel analysis: Techniques and
applications (2nd ed.). New York, NY: Routledge.
}
\keyword{datasets}
