% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormle.R
\name{ormle}
\alias{ormle}
\title{ormle}
\usage{
ormle(est, covmtrx, constr, rhs, nec)
}
\arguments{
\item{est}{PARAM_DESCRIPTION}

\item{covmtrx}{PARAM_DESCRIPTION}

\item{constr}{PARAM_DESCRIPTION}

\item{rhs}{PARAM_DESCRIPTION}

\item{nec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{internal}
