% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_ach.R
\docType{data}
\name{reading_ach}
\alias{reading_ach}
\title{Reading achievement data}
\format{
A data frame with 10320 rows and 5 variables.
}
\usage{
data(reading_ach)
}
\description{
Dataset based on Finch, Bolin, and Kelley (2014, p.32).
}
\details{
\tabular{lll}{
  \strong{school} \tab \code{integer} \tab Clustering variable representing
  the school a given participant was enrolled in\cr
  \strong{gender} \tab \code{factor} \tab Binary factor variable representing
  participants' assigned sex\cr
  \strong{age} \tab \code{integer} \tab Participants' age in months\cr
  \strong{geread} \tab \code{numeric} \tab Reading achievement\cr
  \strong{gevocab} \tab \code{numeric} \tab Vocabulary
}
}
\references{
Finch, W. H., Bolin, J. E., & Kelley, K. (2014). Multilevel
modeling using r. CRC Press 2014.
}
\keyword{datasets}
