% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stevens_1999.R
\docType{data}
\name{stevens_1999}
\alias{stevens_1999}
\title{Sesame Street data based on Stevens (1999)}
\format{
A data frame with 240 rows and 14 variables.
}
\usage{
data(stevens_1999)
}
\description{
Synthetic data based Stevens (1999, p. 596). This study
evaluates the effects of the first year of the Sesame Street television series
in a sample of 3-5 years old children in the USA (N = 240).
}
\details{
\tabular{lll}{
  \strong{age} \tab \code{numeric} \tab Age in months\cr
  \strong{prebody} \tab \code{numeric} \tab Pretest on knowledge of
  body parts\cr
  \strong{prelet} \tab \code{numeric} \tab Pretest on knowledge of letters\cr
  \strong{preform} \tab \code{numeric} \tab Pretest on knowledge of forms\cr
  \strong{prenumb} \tab \code{numeric} \tab Pretest on knowledge of
  numbers\cr
  \strong{prerelat} \tab \code{numeric} \tab Pretest on knowledge of
  relational terms\cr
  \strong{preclas} \tab \code{numeric} \tab Pretest on classification
  skills\cr
  \strong{postbody} \tab \code{numeric} \tab Posttest on knowledge of
  body parts\cr
  \strong{postlet} \tab \code{numeric} \tab Posttest on knowledge of letters\cr
  \strong{postform} \tab \code{numeric} \tab Posttest on knowledge of forms\cr
  \strong{postnumb} \tab \code{numeric} \tab Posttest on knowledge of
  numbers\cr
  \strong{postrelat} \tab \code{numeric} \tab Posttest on knowledge of
  relational terms\cr
  \strong{postclas} \tab \code{numeric} \tab Posttest on classification
  skills\cr
  \strong{peabody} \tab \code{numeric} \tab Mental age score obtained from
  the Peabody Picture Vocabulary test
}
}
\references{
Stevens, J. (1999). Applied multivariate statistics for the
social sciences. (3rd ed.). New Jersey, Lawrance Erlbaum Associates, Inc.
}
\keyword{datasets}
