% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hypotheses.R
\name{compare_hypotheses}
\alias{compare_hypotheses}
\title{Evaluate Informative Hypotheses}
\usage{
compare_hypotheses(object, ..., iter = 1e+05)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{iter}{PARAM_DESCRIPTION, Default: 1e+05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
  #Model fitting
  model <- glm(num_awards ~ -1+prog + math + prog * math, family = "poisson",
               data = academic_awards)
  new_gor <- gorica(x = model, hypothesis = "math=0&math+progGeneral:math=0&math+progVocational:math=0;progGeneral:math > progVocational:math&progVocational:math>0; progVocational:math<0")
}
}
