% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorica.R
\name{compare_hypotheses}
\alias{compare_hypotheses}
\title{Evaluate Informative Hypotheses}
\usage{
compare_hypotheses(object, ..., iter = 1e+05)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{iter}{PARAM_DESCRIPTION, Default: 1e+05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
