% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.R
\name{get_estimates.mplus.model}
\alias{get_estimates.mplus.model}
\title{library(motley)
tmp <- readModels("c:/git_repositories/tmp2_s.out")
gorica(readModels("c:/git_repositories/tmp2_s.out"), "Intercepts.M_3.H>Intercepts.DE_4.H")
tmp3 <- readModels("C:/Git_Repositories/conflict_spillover/mplus/16-04-2018_ang_m_dsem.out")
par_spec <- tmp3$tech1$parameterSpecification
tmp <- get_estimates(readModels("binary/ex9.11.out"))
tmp <- get_estimates(readModels("binary/ex3.12.out"))
tmp5 <- get_estimates(readModels("c:/git_repositories/tmp4.out"))
tmp5$
x <- tmp5
par_spec <- tmp4$tech1$parameterSpecification
estimate <- tmp4$parameters$unstandardized}
\usage{
\method{get_estimates}{mplus.model}(x, ...)
}
\description{
library(motley)
tmp <- readModels("c:/git_repositories/tmp2_s.out")
gorica(readModels("c:/git_repositories/tmp2_s.out"), "Intercepts.M_3.H>Intercepts.DE_4.H")
tmp3 <- readModels("C:/Git_Repositories/conflict_spillover/mplus/16-04-2018_ang_m_dsem.out")
par_spec <- tmp3$tech1$parameterSpecification
tmp <- get_estimates(readModels("binary/ex9.11.out"))
tmp <- get_estimates(readModels("binary/ex3.12.out"))
tmp5 <- get_estimates(readModels("c:/git_repositories/tmp4.out"))
tmp5$
x <- tmp5
par_spec <- tmp4$tech1$parameterSpecification
estimate <- tmp4$parameters$unstandardized
}
