% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorica_methods.R
\name{gorica}
\alias{gorica}
\title{Bayes factors for informative hypotheses}
\usage{
gorica(x, hypothesis, ...)
}
\arguments{
\item{x}{An R object containing the outcome of a statistical analysis.
Currently, the following objects can be processed:
\itemize{
\item \code{lm()} objects (anova, ancova, multiple regression). In the
Example section it is elaborated which calls to \code{lm} can be processed by
\code{gorica} (other calls cannot be processed).
\item \code{t.test()} objects (Student's t-test, Welch's t-test, paired
samples t-test, one-group t-test, equivalence test).
\item A named vector containing the estimates resulting from a statistical
analysis.
Note that, named means that each estimate has to be labelled such that it can
be referred to in \code{hypotheses}.
}}

\item{hypothesis}{A character string containing the informative hypotheses to
evaluate (see Details).}

\item{...}{Additional arguments (see Details).}
}
\value{
Bla
}
\description{
\code{gorica} is an acronym for "Bayesian informative hypotheses evaluation".
It uses the Bayes factor to evaluate hypotheses specified using equality and
inequality constraints among (linear combinations of) parameters in a wide
range of statistical models. A tutorial is provided by Hoijtink, Mulder,
van Lissa, and Gu (2018) retrievable from the Psychological Methods website
at \url{https://www.apa.org/pubs/journals/met/} or the gorica website at
\url{https://informative-hypotheses.sites.uu.nl/software/gorica/} Users are
advised to read this tutorial before using \code{gorica}.
}
\details{
\code{hypotheses} is a character string that specifies which informative
hypotheses have to be evaluated. A simple example is \code{hypotheses <- "a >
b > c; a = b = c;"} which specifies two hypotheses using three estimates with
names "a", "b", and "c", respectively.

The hypotheses specified have to adhere to the following rules:
\enumerate{
\item Parameters are referred to using the names specified in \code{names()}.
\item Linear combinations of parameters must be specified adhering to the
following rules:
        \enumerate{ \item Each parameter name is used at most once.
                    \item Each parameter name may or may not be
                    pre-multiplied with a number.
                    \item A constant may be added or subtracted from each
                    parameter name.
                    \item A linear combination can also be a single number.}
Examples are: \code{3 * a + 5}; \code{a + 2 * b + 3 * c - 2}; \code{a - b};
and \code{5}.
\item (Linear combinations of) parameters can be constrained using <, >, and
=. For example, \code{a > 0} or
\code{a > b = 0} or \code{2 * a < b + c > 5}.
\item The ampersand & can be used to combine different parts of a hypothesis.
For example, \code{a > b & b > c} which is equivalent to \code{a > b > c} or
\code{a > 0 & b > 0 & c > 0}.
\item Sets of (linear combinations of) parameters subjected to the same
constraints can be specified using (). For
example, \code{a > (b,c)} which is equivalent to \code{a > b & a > c}.
\item The specification of a hypothesis is completed by typing ; For example,
\code{hypotheses <- "a > b > c; a = b = c;"}, specifies two hypotheses.
\item Hypotheses have to be compatible, non-redundant and possible. What
these terms mean will be elaborated below.
}

\emph{The set of hypotheses has to be compatible}. For the statistical
background of this requirement see Gu, Mulder, Hoijtink (2018). Usually the
sets of hypotheses specified by researchers are compatible, and if not,
\code{gorica} will return an error message. The following steps can be used to
determine if a set of hypotheses is compatible:
\enumerate{
\item	Replace a range constraint, e.g., \code{1 < a1 < 3}, by an equality
constraint in which the parameter involved is equated to the midpoint of the
range, that is, \code{a1 = 2}.
\item Replace in each hypothesis the < and > by =. For example, \code{a1 = a2
> a3 > a4} becomes \code{a1 = a2 = a3 = a4}.
\item The hypotheses are compatible if there is at least one solution to the
resulting set of equations. For the two hypotheses considered under 1. and
2., the solution is a1 = a2 = a3 = a4 = 2. An example of two non-compatible
hypotheses is \code{hypotheses <- "a = 0; a > 2;"} because there is no
solution to the equations \code{a=0} and \code{a=2}.
}

\emph{Each hypothesis in a set of hypotheses has to be non-redundant.} A
hypothesis is redundant if it can also be specified with fewer constraints.
For example, \code{a = b & a > 0 & b > 0} is redundant because it can also be
specified as \code{a = b & a > 0}. \code{gorica} will work correctly if
hypotheses specified using only < and > are redundant. \code{gorica} will
return an error message if hypotheses specified using at least one = are
redundant.

\emph{Each hypothesis in a set of hypotheses has to be possible.} An
hypothesis is impossible if estimates in agreement with the hypothesis do not
exist. For example: values for \code{a} in agreement with \code{a = 0 &
a > 2} do not exist. It is the responsibility of the user to ensure that the
hypotheses specified are possible. If not, \code{gorica} will either return an
error message or render an output table containing \code{Inf}'s.
}
\examples{
print("bla")
}
\references{
All references are retrievable via or from
\url{https://informative-hypotheses.sites.uu.nl/software/gorica/}
}
\seealso{
The gorica website at
}
\author{
Caspar van Lissa, Yassin Altinisik, Rebecca Kuiper
}
